+++ test/healthcheck.test.js
@@ -24,19 +24,18 @@ test('healthcheck with all defaults does not return an error, but a good respons
   const fastify = Fastify()
   fastify.register(require('../')) // configure this plugin with its default options
 
-  fastify.listen(0, (err) => {
+  fastify.listen(0, (err, address) => {
     fastify.server.unref()
     t.error(err)
-    const port = fastify.server.address().port
 
     sget({
       method: 'GET',
-      timeout: 2000,
-      url: `http://localhost:${port}/health`
+      timeout: 5000,
+      url: `${address}/health`
     }, (err, response, body) => {
       t.error(err)
       t.strictEqual(response.statusCode, 200)
-      t.strictEqual(response.headers['content-type'], 'application/json')
+      t.strictEqual(response.headers['content-type'], 'application/json; charset=utf-8')
       t.deepEqual(JSON.parse(body), { statusCode: 200, status: 'UP' })
 
       fastify.close()
@@ -51,19 +50,18 @@ test('healthcheck on a custom route does not return an error, but a good respons
     'url': '/custom-health'
   }) // configure this plugin with some custom options
 
-  fastify.listen(0, (err) => {
+  fastify.listen(0, (err, address) => {
     fastify.server.unref()
     t.error(err)
-    const port = fastify.server.address().port
 
     sget({
       method: 'GET',
-      timeout: 2000,
-      url: `http://localhost:${port}/custom-health`
+      timeout: 5000,
+      url: `${address}/custom-health`
     }, (err, response, body) => {
       t.error(err)
       t.strictEqual(response.statusCode, 200)
-      t.strictEqual(response.headers['content-type'], 'application/json')
+      t.strictEqual(response.headers['content-type'], 'application/json; charset=utf-8')
       t.deepEqual(JSON.parse(body), { statusCode: 200, status: 'UP' })
 
       fastify.close()
@@ -72,16 +70,16 @@ test('healthcheck on a custom route does not return an error, but a good respons
     // ensure default url is not exposed
     sget({
       method: 'GET',
-      timeout: 2000,
-      url: `http://localhost:${port}/health`
+      timeout: 5000,
+      url: `${address}/health`
     }, (err, response, body) => {
       t.error(err)
       t.strictEqual(response.statusCode, 404)
-      t.strictEqual(response.headers['content-type'], 'application/json')
+      t.strictEqual(response.headers['content-type'], 'application/json; charset=utf-8')
       // t.strictEqual(JSON.parse(body), { statusCode: 404, error: 'Not Found', message: 'Not found' })
       t.strictEqual(JSON.parse(body).statusCode, 404)
       t.strictEqual(JSON.parse(body).error, 'Not Found')
-      t.strictEqual(JSON.parse(body).message, 'Not found')
+      t.strictEqual(JSON.parse(body).message, 'Not Found')
 
       fastify.close()
     })
@@ -96,22 +94,21 @@ test('healthcheck on a disabled route (default or custom), return a not found er
     healthcheckUrlDisable: true
   }) // configure this plugin with some custom options
 
-  fastify.listen(0, (err) => {
+  fastify.listen(0, (err, address) => {
     fastify.server.unref()
     t.error(err)
-    const port = fastify.server.address().port
 
     sget({
       method: 'GET',
-      timeout: 2000,
-      url: `http://localhost:${port}/custom-health`
+      timeout: 5000,
+      url: `${address}/custom-health`
     }, (err, response, body) => {
       t.error(err)
       t.strictEqual(response.statusCode, 404)
-      t.strictEqual(response.headers['content-type'], 'application/json')
+      t.strictEqual(response.headers['content-type'], 'application/json; charset=utf-8')
       t.strictEqual(JSON.parse(body).statusCode, 404)
       t.strictEqual(JSON.parse(body).error, 'Not Found')
-      t.strictEqual(JSON.parse(body).message, 'Not found')
+      t.strictEqual(JSON.parse(body).message, 'Not Found')
 
       fastify.close()
     })
@@ -119,15 +116,15 @@ test('healthcheck on a disabled route (default or custom), return a not found er
     // ensure default url is not exposed
     sget({
       method: 'GET',
-      timeout: 2000,
-      url: `http://localhost:${port}/health`
+      timeout: 5000,
+      url: `${address}/health`
     }, (err, response, body) => {
       t.error(err)
       t.strictEqual(response.statusCode, 404)
-      t.strictEqual(response.headers['content-type'], 'application/json')
+      t.strictEqual(response.headers['content-type'], 'application/json; charset=utf-8')
       t.strictEqual(JSON.parse(body).statusCode, 404)
       t.strictEqual(JSON.parse(body).error, 'Not Found')
-      t.strictEqual(JSON.parse(body).message, 'Not found')
+      t.strictEqual(JSON.parse(body).message, 'Not Found')
 
       fastify.close()
     })
@@ -142,19 +139,18 @@ test('healthcheck with always failure flag, always return an error, (500) and so
     healthcheckUrlAlwaysFail: true
   }) // configure this plugin with some custom options
 
-  fastify.listen(0, (err) => {
+  fastify.listen(0, (err, address) => {
     fastify.server.unref()
     t.error(err)
-    const port = fastify.server.address().port
 
     sget({
       method: 'GET',
-      timeout: 2000,
-      url: `http://localhost:${port}/health`
+      timeout: 5000,
+      url: `${address}/health`
     }, (err, response, body) => {
       t.error(err)
       t.strictEqual(response.statusCode, 500)
-      t.strictEqual(response.headers['content-type'], 'application/json')
+      t.strictEqual(response.headers['content-type'], 'application/json; charset=utf-8')
       t.deepEqual(JSON.parse(body), { statusCode: 500, status: 'DOWN' })
 
       fastify.close()